requirements.txt
asgiref==3.8.1
black==25.1.0
click==8.2.1
crispy-bootstrap5==2025.4
Django==5.2.1
django-crispy-forms==2.4
django-extensions==4.1
django-filter==25.1
django-multiselectfield==0.1.13
django-phonenumber-field==8.1.0
djangorestframework==3.16.0
drf-yasg==1.21.10
Faker==37.3.0
flake8==7.2.0
inflection==0.5.1
isort==6.0.1
Markdown==3.8
marshmallow==4.0.0
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
phonenumbers==9.0.5
pillow==11.2.1
platformdirs==4.3.8
pycodestyle==2.13.0
pydot==4.0.0
pyflakes==3.3.2
pyparsing==3.2.3
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
webargs==8.7.0
